Use Android Studio instead of Eclipse.
