+++ app/.settings/gradle/org.springsource.ide.eclipse.gradle.core.prefs
@@ -1,0 +1,4 @@
+#org.springsource.ide.eclipse.gradle.core.preferences.GradleProjectPreferences
+#Sun Apr 26 17:39:03 CEST 2015
+org.springsource.ide.eclipse.gradle.linkedresources=
+org.springsource.ide.eclipse.gradle.rootprojectloc=..
